Base Produktion und mehr

+++ Assets/Scripts/ScriptableObjectClasses/StructureScriptableObject.cs
@@ -8,7 +8,7 @@ public class StructureScriptableObject : ScriptableObject
     public string Name;
     public StructureType Type;
     public List<Resource> Resources;
-    public float EcoImpactFactor;
+    public int EcoImpactFactor;
     public List<AllowedLocation> AllowedLocations;
     public GameObject Symbol;
 }
